backend/requirements.txt
django
Pillow
djangorestframework
markdown
django-filter
djangorestframework-simplejwt
django-cors-headers